playground/webclient/src/test/java/com/example/webclient/ReviewClientTest.java
package com.example.webclient;

import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
import static com.github.tomakehurst.wiremock.client.WireMock.get;
import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
import static org.assertj.core.api.Assertions.assertThat;
import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;

import com.github.tomakehurst.wiremock.client.WireMock;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import reactor.test.StepVerifier;

@SpringBootTest(webEnvironment = RANDOM_PORT)
@AutoConfigureWireMock(port = 9090)
class ReviewClientTest {

  @Autowired
  ReviewClient reviewClient;

  @Test

  void getReviews() {
    long productId = 1;
    stubFor(get(urlPathEqualTo("/v1/reviews"))
        .willReturn(aResponse()
            .withStatus(HttpStatus.OK.value())
            .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
            .withBodyFile("reviews.json")
        ));


    StepVerifier.create(reviewClient.getReviews(productId))
        .assertNext(reviewDtos -> {
          assertThat(reviewDtos).hasSize(2);
        })
        .verifyComplete();

    WireMock.verify(1, getRequestedFor(urlPathEqualTo("/v1/reviews")));
  }
}